+++ R/IF_ESRatio.R
@@ -33,7 +33,7 @@
 #'                     retVals = NULL, nuisPars = NULL,
 #'                     IFplot = TRUE, IFprint = TRUE)
 #'
-#' data(edhec)
+#' data(edhec, package = "PerformanceAnalytics")
 #' colnames(edhec) = c("CA", "CTAG", "DIS", "EM","EMN", "ED", "FIA",
 #'                     "GM", "LS", "MA", "RV", "SS", "FoF") 
 #'