+++ R/IF_mean.R
@@ -1,6 +1,6 @@
 #' @title Influence Function - Mean
 #'
-#' @description \code{IF.mean} returns the data and plots the shape of either the IF or the IF TS for the mean.
+#' @description \code{IF.Mean} returns the data and plots the shape of either the IF or the IF TS for the mean.
 #'
 #' @param returns Returns data of the asset or portfolio. This can be a numeric or an xts object.
 #' @param evalShape Evaluation of the shape of the IF risk or performance measure if TRUE. Otherwise, a TS of the IF of the provided returns is computed.
@@ -27,26 +27,26 @@
 #' 
 #' @examples
 #' # Plot of IF with nuisance parameter with return value
-#' outIF <- IF.mean(returns = NULL, evalShape = TRUE, retVals = NULL, nuisPars = NULL,
+#' outIF <- IF.Mean(returns = NULL, evalShape = TRUE, retVals = NULL, nuisPars = NULL,
 #'                  IFplot = TRUE, IFprint = TRUE)
 #'
-#' data(edhec)
+#' data(edhec, package = "PerformanceAnalytics")
 #' colnames(edhec) = c("CA", "CTAG", "DIS", "EM","EMN", "ED", "FIA",
 #'                     "GM", "LS", "MA", "RV", "SS", "FoF") 
 #' 
 #' # Plot of IF a specified TS 
-#' outIF <- IF.mean(estimator = "mean",
+#' outIF <- IF.Mean(estimator = "mean",
 #'                  returns = edhec[,"CA"], evalShape = TRUE, 
 #'                  retVals = seq(-0.1, 0.1, by = 0.001), nuisPars = NULL,
 #'                  IFplot = TRUE, IFprint = TRUE)
 #' 
 #' # Computing the IF of the returns (with prewhitening) with a plot of IF TS
-#' outIF <- IF.mean(returns = edhec[,"CA"], evalShape = FALSE, 
+#' outIF <- IF.Mean(returns = edhec[,"CA"], evalShape = FALSE, 
 #'                  retVals = NULL, nuisPars = NULL,
 #'                  IFplot = TRUE, IFprint = TRUE,
 #'                  prewhiten = FALSE)
 #'
-IF.mean <- function(returns = NULL, evalShape = FALSE, retVals = NULL, nuisPars = NULL, k = 4,
+IF.Mean <- function(returns = NULL, evalShape = FALSE, retVals = NULL, nuisPars = NULL, k = 4,
                     IFplot = FALSE, IFprint = TRUE, 
                     prewhiten = FALSE, ar.prewhiten.order = 1,
                     cleanOutliers = FALSE, cleanMethod = c("locScaleRob")[1], eff = 0.99, 
@@ -75,7 +75,7 @@ IF.mean <- function(returns = NULL, evalShape = FALSE, retVals = NULL, nuisPars 
   
   # Plot for shape evaluation
   if(evalShape){
-    IFvals <- EvaluateShape(estimator = "mean",
+    IFvals <- EvaluateShape(estimator = "Mean",
                             retVals = retVals, returns = returns, k = k, nuisPars = nuisPars,
                             IFplot = IFplot, IFprint = IFprint)
     if(IFprint)
@@ -87,19 +87,19 @@ IF.mean <- function(returns = NULL, evalShape = FALSE, retVals = NULL, nuisPars 
   }
 
   # Computing the IF vector for the mean of the returns
-  IF.mean.vector <- IF.mean.fn(x = returns, returns = returns)
+  IF.Mean.vector <- IF.Mean.fn(x = returns, returns = returns)
   
   # Adding the pre-whitening functionality  
   if(prewhiten)
-    IF.mean.vector <- as.numeric(arima(x = IF.mean.vector, order = c(ar.prewhiten.order,0,0), include.mean = TRUE)$residuals)
+    IF.Mean.vector <- as.numeric(arima(x = IF.Mean.vector, order = c(ar.prewhiten.order,0,0), include.mean = TRUE)$residuals)
   
   # Adjustment for data (xts)
   if(xts::is.xts(returns))
-    IF.mean.vector <- xts::xts(IF.mean.vector, returns.dates)
+    IF.Mean.vector <- xts::xts(IF.Mean.vector, returns.dates)
   
   # Plot of the IF TS
   if(isTRUE(IFplot)){
-    print(plot(IF.mean.vector, type = "l", main = "Mean Estimator Influence Function Transformed Returns", ylab = "IF"))
+    print(plot(IF.Mean.vector, type = "l", main = "Mean Estimator Influence Function Transformed Returns", ylab = "IF"))
   }
   
   # Stop if no printing of the TS
@@ -110,5 +110,5 @@ IF.mean <- function(returns = NULL, evalShape = FALSE, retVals = NULL, nuisPars 
   }
   
   # Returning the IF vector for the mean of the returns
-  return(IF.mean.vector)
+  return(IF.Mean.vector)
 }