+++ R/IF_robMean.R
@@ -25,7 +25,7 @@
 #' @export
 #'
 #' @examples
-#' data(edhec)
+#' data(edhec, package = "PerformanceAnalytics")
 #' colnames(edhec) = c("CA", "CTAG", "DIS", "EM","EMN", "ED", "FIA",
 #'                     "GM", "LS", "MA", "RV", "SS", "FoF") 
 #'